+++ CHANGES.md
@@ -18,7 +18,7 @@ Features:
 -
 
 Changes:
-- 
+-
 
 Fixes:
 - fix `okla` improve cleanup_content method to use CSS selectors and return bytes #1692