+++ CHANGES.md
@@ -30,6 +30,7 @@ Fixes:
 
 Fixes:
 - fix recap.email docket_number parsing inside nested tags
+- fix `uscgcoca` and `asbca` headers #1664
 
 ## Past