+++ CHANGES.md
@@ -29,6 +29,7 @@ Fixes:
 
 Fixes:
 - fix recap.email docket_number parsing inside nested tags
+- fix `uscgcoca` and `asbca` headers #1664
 
 ## Past