+++ CHANGES.md
@@ -22,7 +22,8 @@ Changes:
 
 Fixes:
 - fix `okla` improve cleanup_content method to use CSS selectors and return bytes #1692
-
+- Fix `fla` and `fladistctapp` scrapers by scraping new API endpoint #1690
+- fix `uscgcoca` and `asbca` headers #1664
 
 ## Current